+++ Nickel.Bugfixes/GameEnums.cs
@@ -1,0 +1,6 @@
+﻿using Nanoray.EnumByNameSourceGenerator;
+
+namespace Nickel.Bugfixes;
+
+[EnumByName(typeof(Spr))]
+internal static partial class StableSpr;