+++ Nickel.Bugfixes/Fixes/EnergyFragmentFixes.cs
@@ -1,0 +1,54 @@
+﻿using HarmonyLib;
+using Microsoft.Extensions.Logging;
+using Nanoray.Shrike;
+using Nanoray.Shrike.Harmony;
+using System;
+using System.Collections.Generic;
+using System.Reflection;
+using System.Reflection.Emit;
+
+namespace Nickel.Bugfixes;
+
+internal static class EnergyFragmentFixes
+{
+	public static void ApplyPatches(IHarmony harmony)
+		=> harmony.Patch(
+			original: AccessTools.DeclaredMethod(typeof(AStatus), nameof(AStatus.Begin))
+					?? throw new InvalidOperationException($"Could not patch game methods: missing method `{nameof(AStatus)}.{nameof(AStatus.Begin)}`"),
+			transpiler: new HarmonyMethod(MethodBase.GetCurrentMethod()!.DeclaringType!, nameof(AStatus_Begin_Transpiler))
+		);
+
+	private static IEnumerable<CodeInstruction> AStatus_Begin_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod, ILGenerator il)
+	{
+		// ReSharper disable PossibleMultipleEnumeration
+		try
+		{
+			return new SequenceBlockMatcher<CodeInstruction>(instructions)
+				.Find([
+					ILMatches.Ldarg(0).CreateLabel(il, out var ifLabel),
+					ILMatches.Ldfld("status"),
+					ILMatches.LdcI4((int)Status.energyFragment),
+					ILMatches.BneUn,
+				])
+				.Find([
+					ILMatches.Ldarg(3),
+					ILMatches.Instruction(OpCodes.Dup),
+					ILMatches.Ldfld("energy"),
+					ILMatches.LdcI4(1),
+					ILMatches.Instruction(OpCodes.Add),
+					ILMatches.Stfld("energy"),
+				])
+				.Insert(
+					SequenceMatcherPastBoundsDirection.After, SequenceMatcherInsertionResultingBounds.IncludingInsertion,
+					new CodeInstruction(OpCodes.Br, ifLabel)
+				)
+				.AllElements();
+		}
+		catch (Exception ex)
+		{
+			ModEntry.Instance.Logger.LogError("Could not patch method {Method} - {Mod} probably won't work.\nReason: {Exception}", originalMethod, ModEntry.Instance.Package.Manifest.UniqueName, ex);
+			return instructions;
+		}
+		// ReSharper restore PossibleMultipleEnumeration
+	}
+}